UCBQuizMaster/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0

